+++ src/features/upload/components/upload-guide.tsx
